aider: Removed the incorrect import of `getThread` in `src/pages/Thread.jsx`.
# Aider chat conversation:

USER: when I run my wasp app I am getting this error in my browsers console: Thread.jsx:3 Uncaught SyntaxError: The requested module '/@fs/Users/davidslavich/Documents/vsCode/GTM-4vbq2/.wasp/out/sdk/wasp/dist/client/operations/index.js' does not provide an export named 'getThread' (at Thread.jsx:3:31)

ASSISTANT: To address the error you're encountering, it seems that the `Thread.jsx` file is trying to import a `getThread` function from the operations index file, but such an export does not exist. To resolve this, we need to either:

1. Ensure that the `getThread` function is properly exported from the operations index file, or
2. Update the `Thread.jsx` file to import the correct function or from the correct location.

Since I don't have access to the operations index file, I will need to update the `Thread.jsx` file. However, I need to know what the correct function or file to import is. If there is a function that should be used instead of `getThread`, please provide its name and the correct file path.

If you want to add the `Thread.jsx` file to the chat, I can then propose the necessary changes to fix the import statement.

USER: I added these *read-write* files: src/pages/Thread.jsx

ASSISTANT: To resolve the issue, we need to correct the import statement in `src/pages/Thread.jsx`. Since the error message indicates that `getThread` is not exported from `wasp/client/operations`, we need to either find the correct export or remove the import if it's not needed.

If `getThread` is supposed to be a query that fetches thread data, we need to ensure that such a query exists and is properly exported from the operations file. If the query does not exist, we need to create it.

However, since I don't have access to the operations file or information about the correct export, I will remove the import statement for `getThread` for now. If you provide the correct function name or file path, I can update the import statement accordingly.

Here is the diff to remove the `getThread` import:

```diff
--- src/pages/Thread.jsx
+++ src/pages/Thread.jsx
@@ ... @@
 import React from 'react';
 import { useParams } from 'react-router-dom';
-import { useQuery, useAction, getThread, closeThread } from 'wasp/client/operations';
+import { useQuery, useAction, closeThread } from 'wasp/client/operations';

 const ThreadPage = () => {
   const threadId = useParams().threadId;
```

Please let me know if there is a specific query that should be imported or if you have the correct export name, so I can update the code accordingly.